Use reversed CSV file; various fixes
Inventories work if buy transaction is booked before sell transaction.

Degiro provides CSV ordered by time descending, so a sell appears
first.

This will not be fixed later by Beancount, because it orders the
transactions by date (granularity: 1 day) and by metadata line number.

However, if a stock is bought and sold on the same day, Beancount puts
them in the false order, because the sell transaction has a lower line
number.

The only feasible solution is to reverse the order of entries in the
input file before importing, e.g.:

{ head -n 1 Account.csv && tac Account.csv | head -n -1 }  > Account_reversed.csv